Add filesystem_store_metadata_file parameter
This parameter conatins a path to a valid file which contains the
metadata to be returned with any location associated with the filesystem
store.

While enabling glance netapp nfs, this parameters needs to be set with
file path which must contain a valid JSON object with the keys 'id' and
'mountpoint'.